Add multiple MediatR pipeline behavior
This commit adds one logging behavior for commands and queries and another one that disables Entity Framework Core's tracking during the execution of queries.
The command/query logging behaviors log the respective command/query.
In order to avoid duplication of logging a custom exception (LoggedMediatRException) is thrown by the behaviors to notify the global handler that it can skip the logging part.